+++ test/test_smiley.jl
@@ -1,33 +1,23 @@
 include("../examples/smiley_examples.jl")
 
-using Test
-using IntervalArithmetic, IntervalRootFinding
 using .SmileyExample22, .SmileyExample52, .SmileyExample54, .SmileyExample55
 
-function test_all_unique(xs)
-    for x in xs
-        @test x.status == :unique
-    end
-    return nothing
-end
-
 const abstol = 1e-6
-for contractor in (Newton, Krawczyk) # NOTE: Bisection method performs badly in all examples
-
-@info("Testing method $contractor")
 
 @testset "$(SmileyExample22.title)" begin
-    roots_found = roots(SmileyExample22.f, SmileyExample22.region ; contractor, abstol)
-    @test length(roots_found) == 8
-    test_all_unique(roots_found)
-    # no reference data for roots given
+    for contractor in (Newton, Krawczyk)
+        roots_found = roots(SmileyExample22.f, SmileyExample22.region ; contractor, abstol)
+        @test length(roots_found) == 8
+        @test all(isunique, roots_found)
+        # no reference data for roots given
+    end
 end
 
-for example in (SmileyExample52, SmileyExample54)#, SmileyExample55)
-    @testset "$(example.title)" begin
+@testset "$(example.title)" for example in (SmileyExample52, SmileyExample54)
+    for contractor in (Newton, Krawczyk)
         roots_found = roots(example.f, example.region ; contractor, abstol)
         @test length(roots_found) == length(example.known_roots)
-        test_all_unique(roots_found)
+        @test all(isunique, roots_found)
         for rf in roots_found
             # check there is exactly one known root for each found root
             @test 1 == sum(example.known_roots) do rk
@@ -36,5 +26,3 @@ for example in (SmileyExample52, SmileyExample54)#, SmileyExample55)
         end
     end
 end
-
-end # method loop
